Moved type validation to framework
