+++ pkg/epp/framework/interface/datalayer/plugin.go
@@ -18,6 +18,8 @@ package datalayer
 
 import (
 	"context"
+	"errors"
+	"fmt"
 	"reflect"
 
 	"sigs.k8s.io/gateway-api-inference-extension/pkg/epp/framework/interface/plugin"
@@ -49,4 +51,20 @@ type Extractor interface {
 	// Extract transforms the raw data source output into a concrete structured
 	// attribute, stored on the given endpoint.
 	Extract(ctx context.Context, data any, ep Endpoint) error
+}
+
+// ValidateExtractorType checks if an extractor can handle
+// the DataSource's output. It should be called by a DataSource
+// when an extractor is added.
+func ValidateExtractorType(collectorOutput, extractorInput reflect.Type) error {
+	if collectorOutput == nil || extractorInput == nil {
+		return errors.New("extractor input type or data source output type can't be nil")
+	}
+	if collectorOutput == extractorInput ||
+		(extractorInput.Kind() == reflect.Interface && extractorInput.NumMethod() == 0) ||
+		(extractorInput.Kind() == reflect.Interface && collectorOutput.Implements(extractorInput)) {
+		return nil
+	}
+	return fmt.Errorf("extractor input type %v cannot handle data source output type %v",
+		extractorInput, collectorOutput)
 }
